refactoring aggregated stats to biodata-tools.
added classes and removed some methods from the factories.
