+++ biodata-tools/src/main/java/org/opencb/biodata/tools/variant/stats/VariantAggregatedStatsCalculator.java
@@ -1,0 +1,9 @@
+package org.opencb.biodata.tools.variant.stats;
+
+/**
+ * Created by jmmut on 2015-08-25.
+ *
+ * @author Jose Miguel Mut Lopez &lt;jmmut@ebi.ac.uk&gt;
+ */
+public class VariantAggregatedStatsCalculator {
+}
