+++ biodata-tools/src/test/java/org/opencb/biodata/tools/variant/stats/VariantStatsCalculatorTest.java
@@ -1,0 +1,14 @@
+package org.opencb.biodata.tools.variant.stats;
+
+import org.opencb.commons.test.GenericTest;
+
+import static org.junit.Assert.*;
+
+/**
+ * Created by jmmut on 2015-08-25.
+ *
+ * @author Jose Miguel Mut Lopez &lt;jmmut@ebi.ac.uk&gt;
+ */
+public class VariantStatsCalculatorTest extends GenericTest {
+
+}
